+++ cosmonaut-frontend/src/components/Chapter/lesson4/chapter3/L4C3Pr.jsx
@@ -13,7 +13,7 @@ import Markdown from "../../../../components/Contents/Markdown";
 import { useNavigate, useParams } from "react-router-dom";
 import PracticeCode from "../../../../components/CodeEditor/PracticeCode";
 import { Loading } from "../../../../components/Common/Loading";
-import EditorPr from "../../../../components/CodeEditor/EditorPr";
+import EditorPractice from "../../../../components/CodeEditor/EditorPractice";
 import ResultTab from "../../../../components/CodeEditor/ResultTab";
 import TabHeader from "../../../../components/Practice/TabHeader";
 import PracticeName from "../../../../components/Practice/PracticeName";
@@ -298,7 +298,7 @@ export const L4C3Pr = () => {
                     <Loading />
                   ) : (
                     <>
-                      <EditorPr
+                      <EditorPractice
                         defaultLanguage="rust"
                         exCode={example[tab]}
                         path={key}
