+++ cosmonaut-frontend/src/core/state/loginState.js
@@ -8,6 +8,6 @@ const { persistAtom } = recoilPersist({
 
 export const LoginState = atom({
   key: "LoginState",
-  default: false,
+  default: true,
   effects_UNSTABLE: [persistAtom],
 });
